Rename RMSNorm operator module
