+++ mstar/engine/resources/rms_norm/__init__.py
@@ -1,5 +1,5 @@
 """Backend-neutral RMSNorm operator."""
 
-from mstar.engine.resources.rms_norm.ops import run_rms_norm
+from mstar.engine.resources.rms_norm.rms_norm import run_rms_norm
 
 __all__ = ["run_rms_norm"]
